cost-optimization/gke-vpa-recommendations/gke_metrics_pipeline/2-cloud-monitoring-to-pubsub/requirements.txt
google-cloud-pubsub
google-cloud-bigquery
httpx
pyarrow
pandas_gbq